add some options

+++ package_synchronizer/package_synchronizer.py
@@ -4,6 +4,8 @@ import yaml
 import os
 import hashlib
 import logging
+from logger.logger_extension import LogRecordExtension, Extension
+import datetime
 
 class PackageSynchronizer:
     def __init__(self, config_path):
@@ -24,7 +26,8 @@ class PackageSynchronizer:
         checksum = md5.hexdigest()
         return checksum
 
-    def synchronize(self):
+    def synchronize(self, false_update = False):
+        result = {}
         bucket = boto3.resource('s3').Bucket(self.bucket_name)
         for s3_object in self.response['Contents']:
             package_path = os.path.join(self.local_package_directory, s3_object['Key'])
@@ -32,20 +35,34 @@ class PackageSynchronizer:
             md5_hash = metadata["ResponseMetadata"]["HTTPHeaders"]["x-amz-meta-codebuild-content-md5"]
             if not os.path.exists(os.path.dirname(package_path)):
                 os.makedirs(os.path.dirname(package_path))
-            if not os.path.exists(package_path):
+            action = {}
+            stamp = {}
+            if false_update:
                 bucket.download_file(s3_object['Key'], package_path)
-                print("downloading => " + s3_object['Key'])
-            else :
-                if self.getHash(package_path) != md5_hash :
-                    os.remove(package_path)
+                action = {'action' : 'download new package'}
+                stamp = {'stamp' : datetime.datetime.now().isoformat()}
+            else:
+                if not os.path.exists(package_path):
                     bucket.download_file(s3_object['Key'], package_path)
-                    print("downloading => " + s3_object['Key'])
-                else:
-                    print("skip downloading => " + s3_object['Key'])
+                    action = {'action' : 'download new package'}
+                    stamp = {'stamp' : datetime.datetime.now().isoformat()}
+                else :
+                    if self.getHash(package_path) != md5_hash :
+                        os.remove(package_path)
+                        bucket.download_file(s3_object['Key'], package_path)
+                        action = {'action' : 'update package'}
+                        stamp = {'stamp' : datetime.datetime.now().isoformat()}
+                    else:
+                        action = {'action' : 'skip download package'}
+                        stamp = {'stamp' : datetime.datetime.now().isoformat()}
+            result[s3_object['Key']] = {}
+            result[s3_object['Key']].update(action)
+            result[s3_object['Key']].update(stamp)
+        return result
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='synchronize debian package in S3')
     parser.add_argument('config_path', help='path to the config yaml path')
     args = parser.parse_args()
     sync = PackageSynchronizer(args.config_path)
-    sync.synchronize()
+    sync.synchronize(False)
